requirements.txt
playwright
pandas
fake-useragent
python-dotenv
requests
colorama
tqdm
fastapi
uvicorn
supabase
httpx
openai
